reorganized vacancies page components
